# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BrandOps API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List all campaigns
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, active, paused, completed, archived]
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a new campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Created campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get a campaign by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          description: Not found
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignUpdate"
      responses:
        "200":
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          description: Not found
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /campaigns/{id}/publish:
    post:
      operationId: publishCampaign
      tags: [campaigns]
      summary: Publish a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Published campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}/stats:
    get:
      operationId: getCampaignStats
      tags: [campaigns]
      summary: Get campaign-level statistics
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignStats"
  /creators:
    get:
      operationId: listCreators
      tags: [creators]
      summary: List all creators
      parameters:
        - name: platform
          in: query
          schema:
            type: string
            enum: [tiktok, instagram, youtube]
        - name: niche
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of creators
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Creator"
    post:
      operationId: createCreator
      tags: [creators]
      summary: Register a new creator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatorInput"
      responses:
        "201":
          description: Created creator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Creator"
  /creators/{id}:
    get:
      operationId: getCreator
      tags: [creators]
      summary: Get a creator by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Creator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Creator"
        "404":
          description: Not found
    patch:
      operationId: updateCreator
      tags: [creators]
      summary: Update a creator
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatorUpdate"
      responses:
        "200":
          description: Updated creator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Creator"
...
components:
  schemas:
...
    Campaign:
      type: object
      required:
        - id
        - title
        - description
        - totalBudget
        - payoutPerVideo
        - platform
        - niche
        - status
        - deadline
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        totalBudget:
          type: number
        payoutPerVideo:
          type: number
        platform:
          type: string
          enum: [tiktok, instagram, youtube]
        niche:
          type: string
        status:
          type: string
          enum: [draft, active, paused, completed, archived]
        deadline:
          type: string
          format: date-time
        inspirationUrls:
          type: ["string", "null"]
        creatorCount:
          type: integer
        approvedCount:
          type: integer
        pendingCount:
          type: integer
        totalSpent:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CampaignInput:
      type: object
      required:
        - title
        - description
        - totalBudget
        - payoutPerVideo
        - platform
        - niche
        - deadline
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        totalBudget:
          type: number
        payoutPerVideo:
          type: number
        platform:
          type: string
          enum: [tiktok, instagram, youtube]
        niche:
          type: string
        deadline:
          type: string
          format: date-time
        inspirationUrls:
          type: string
    CampaignUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        totalBudget:
          type: number
        payoutPerVideo:
          type: number
        platform:
          type: string
          enum: [tiktok, instagram, youtube]
        niche:
          type: string
        deadline:
          type: string
          format: date-time
        status:
          type: string
          enum: [draft, active, paused, completed, archived]
        inspirationUrls:
          type: string
    CampaignStats:
      type: object
      required:
        - campaignId
        - totalSubmissions
        - approvedSubmissions
        - pendingSubmissions
        - rejectedSubmissions
        - totalSpent
        - budgetRemaining
        - creatorCount
      properties:
        campaignId:
          type: integer
        totalSubmissions:
          type: integer
        approvedSubmissions:
          type: integer
        pendingSubmissions:
          type: integer
        rejectedSubmissions:
          type: integer
        totalSpent:
          type: number
        budgetRemaining:
          type: number
        creatorCount:
          type: integer
    Creator:
      type: object
      required:
        - id
        - name
        - email
        - platform
        - handle
        - niche
        - followerCount
        - engagementRate
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        platform:
          type: string
          enum: [tiktok, instagram, youtube]
        handle:
          type: string
        niche:
          type: string
        followerCount:
          type: integer
        engagementRate:
          type: number
        avatarUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, suspended]
        totalEarnings:
          type: number
        approvedVideos:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreatorInput:
      type: object
      required:
        - name
        - email
        - platform
        - handle
        - niche
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        platform:
          type: string
          enum: [tiktok, instagram, youtube]
        handle:
          type: string
        niche:
          type: string
        followerCount:
          type: integer
        engagementRate:
          type: number
        avatarUrl:
          type: string
    CreatorUpdate:
      type: object
      properties:
        name:
          type: string
        platform:
          type: string
          enum: [tiktok, instagram, youtube]
        handle:
          type: string
        niche:
          type: string
        followerCount:
          type: integer
        engagementRate:
          type: number
        avatarUrl:
          type: string
        status:
          type: string
          enum: [active, inactive, suspended]
...