# lib/api-zod/src/generated/types/creatorInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BrandOps API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreatorInputPlatform } from './creatorInputPlatform';

export interface CreatorInput {
  /** @minLength 1 */
  name: string;
  email: string;
  platform: CreatorInputPlatform;
  handle: string;
  niche: string;
  followerCount?: number;
  engagementRate?: number;
  avatarUrl?: string;
}